Move `shader_executor` out of transformations
