+++ compositor_render/src/shader_executor/error.rs
