+++ compositor_render/src/transformations.rs
@@ -1,6 +1,5 @@
 pub mod builtin;
 pub mod image_renderer;
 pub mod shader;
-pub mod shader_executor;
 pub mod text_renderer;
 pub mod web_renderer;
